+++ compute_pairs.h
@@ -22,16 +22,16 @@ typedef priority_queue<Cube, vector<Cube>, CubeComparator> CubeQue;
 class ComputePairs{
 private:
 	DenseCubicalGrids * dcg;
-	unordered_map<int, int> pivot_column_index;
-	int dim;
+	unordered_map<uint64_t, uint32_t> pivot_column_index;
+	uint8_t dim;
 	vector<WritePairs> *wp;
 	bool print;
 
 public:
 	ComputePairs(DenseCubicalGrids* _dcg, vector<WritePairs> &_wp, const bool _print);
 	void compute_pairs_main(vector<Cube>& ctr, int min_cache_size);
-	void assemble_columns_to_reduce(vector<Cube>& ctr, int _dim);
-	void add_cache(int i, CubeQue &wc, unordered_map<int, CubeQue>& recorded_wc);
+	void assemble_columns_to_reduce(vector<Cube>& ctr, uint8_t _dim);
+	void add_cache(uint32_t i, CubeQue &wc, unordered_map<uint32_t, CubeQue>& recorded_wc);
 	Cube pop_pivot(vector<Cube>& column);
 	Cube get_pivot(vector<Cube>& column);
 	Cube pop_pivot(CubeQue& column);
